Working on adding player system
Working with sql right now, works well, I'm adding logic for checking if doctor and address exists.

If they do, a existing id will be used in the player's table, if not a new one will be created and used.

I have lil problems with doctor's logic but everything else works well so far.