Calendar Overhaul
I decided to overhaul whole calendar system into something much simpler to work with.

At first I wanted to have a nice calendar that looked professional and had all these fancy stuff, found one on dayPilot, but having to add functionality that I want turned out too complex of a challange as for my skills at the moment.

So I decided to make a simpler version by starting with a simple form in which I can switch between types of event easily, and then build up upon that, and perhaps make a calendar, but as for now it'll be displayed as a list.