Update Tweak
You can now leave the fields blank on the update pages.

If you choose to do so, the existing records from the database will be retrieved and used for updating.

This means that if all fields are left blank, the update operation will essentially retrieve the current data from the database and apply it back, resulting in no actual changes being made.