Style Update
I don't really have a design for this website, and as I said in the Graded Unit planning document, the website will look have similar style to the design images.

Design would take a lot of my time away so I just developed my website with a few designs in mind but it just looks too simple, so I decided to use boostrap to makie it look nicer.