Password Criteria Box
I made some changes so that the items in criteria box (<li> and <i>) - will change their colour and icon depending on password's length and if two password match.

If password length is greater than 12 or the two passwords are identical - text will become green and cross icon will be replaced with check mark icon, otherwise it will be red and the check mark icon will be replaced with a cross icon.