Checkbox Update
I have made an update to the player list in which I have replaced ID number with checkboxes, for yet easier controls.

As for now I have only made debug code, but next I plan to elevate this so that it will redirect user depending on which checkbox is ticked, and action selected.